requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blinker==1.8.2
boto3==1.26.97
botocore==1.29.97
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.2.5
cryptography==3.4.8
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.4
distro==1.7.0
distro-info==1.0
Django==4.0
django-ckeditor==6.5.1
django-crispy-forms==1.13.0
django-js-asset==2.0.0
django-search-views==0.3.7
django-storages==1.13.2
djangorestframework==3.14.0
dnspython
eventlet==0.30.2
filelock==3.4.0
Flask==3.0.3
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.2.0
kafka-python==2.0.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
moviepy==1.0.3
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.8.1
numpy==1.22.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
packaging==24.0
pandas==1.3.5
paramiko==2.9.3
pika==1.3.2
Pillow==9.4.0
platformdirs==2.4.0
plotly==5.22.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
pyasn1==0.4.8
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2022.10.31
requests==2.31.0
requests-toolbelt==0.9.1
retrying==1.3.4
roman==3.3
rsa==3.1.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
stdiomask==0.0.6
systemd-python==234
tenacity==8.3.0
textblob==0.17.1
tqdm==4.65.0
typing_extensions==4.11.0
tzdata==2021.5
zipp==3.18.2
